choir(mapperUtils.ts) Consolidate utils files
