+++ packages/data-explorer/tests/unit/mappers/metaFilterMapper.spec.ts
@@ -20,7 +20,6 @@ describe('metaFilterMapper', () => {
     })
     it('create an filter definition from metadata', async () => {
       const resp = await metaFilterMapper.mapMetaToFilters(meta)
-      console.log(JSON.stringify(resp))
       expect(JSON.stringify(resp)).toEqual('{"definition":[{"name":"country","label":"country","type":"checkbox-filter","collapsable":true,"collapsed":false,"maxVisibleOptions":10},{"name":"test","label":"test","type":"range-filter","collapsable":true,"collapsed":false,"max":10,"min":-10,"useSlider":true},{"name":"age_groups","label":"age_groups","type":"checkbox-filter","collapsable":true,"collapsed":false,"maxVisibleOptions":10}],"shown":[]}')
     })
   })
